Implementing DspSettingsServer.
This Server subclass helps manage audio inputs and outputs for a given
module, and allows for permanently setting the count of inputs and
outputs. That helps solve the problem of a microphone input which only
ever has one channel, a envelope follower with only one useful input, or
an ambisonics panner that always has octophonic outputs.